feat(emails): disable from & name editing; universal from & name settings
