+++ includes/class-magic-link.php
@@ -358,8 +358,8 @@ final class Magic_Link {
 			'headers' => [
 				sprintf(
 					'From: %1$s <%2$s>',
-					Reader_Activation::get_from_name(),
-					Reader_Activation::get_from_email()
+					Emails::get_from_name(),
+					Emails::get_from_email()
 				),
 			],
 		];
